+++ src/num/geom/metric/orientation.rs
@@ -20,22 +20,10 @@ use crate::{Distance, Position};
 /// to be normalized in most use cases.
 #[must_use]
 #[repr(transparent)]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Orientation<T, const D: usize> {
     /// The directional components in `D`-dimensional space.
     pub dim: [T; D],
 }
-crate::items! {
-    impl<T, const D: usize> From<[T; D]> for Orientation<T, D> {
-        fn from(dim: [T; D]) -> Self { Self { dim } }
-    }
-    impl<T> From<(T, T)> for Orientation<T, 2> {
-        fn from(dim: (T, T)) -> Self { Self { dim: [dim.0, dim.1] } }
-    }
-    impl<T> From<(T, T, T)> for Orientation<T, 3> {
-        fn from(dim: (T, T, T)) -> Self { Self { dim: [dim.0, dim.1, dim.2] } }
-    }
-    impl<T> From<(T, T, T, T)> for Orientation<T, 4> {
-        fn from(dim: (T, T, T, T)) -> Self { Self { dim: [dim.0, dim.1, dim.2, dim.3] } }
-    }
-}
+
+crate::_impl_metric![common_methods: Orientation];
+crate::_impl_metric![common_traits: Orientation];