+++ src/num/mod.rs
@@ -56,7 +56,7 @@ crate::items! { // structural access: _mods, _pub_mods, _internals, _all, _alway
         // pub use super::symb::_all::*;
     }
     pub(super) mod _internals { #![allow(unused)]
-        pub(crate) use super::{_private::*, rand::_internals::*};
+        pub(crate) use super::{_private::*, geom::_internals::*, rand::_internals::*};
     }
     pub(super) mod _all {
         #[doc(inline)]